hr_management_app/requirements.txt
Flask==2.0.3
SQLAlchemy==1.4.22
pytest==6.2.4
pandas>=1.4.0
openpyxl>=3.0.0
python-docx>=0.8.11
python-dateutil>=2.8.0
rapidfuzz>=3.0.0
rapidfuzz>=3.0.0
scikit-learn>=1.0.0
joblib>=1.1.0